chore: refactor tasks view
