+++ lib/log_flow/widgets/widgets.dart
@@ -4,3 +4,4 @@ export 'comment_tile.dart';
 export 'continue_button.dart';
 export 'mood_picker.dart';
 export 'progress.dart';
+export 'tasks_view.dart';
